Library

Contents

Library

Contents

Functions

Index

General functions

VTU Keywords

IO-Functions

Functions

Index

VTUFileHandler.VTUData

VTUFileHandler.VTUDataField

VTUFileHandler.VTUFile

VTUFileHandler.VTUHeader

VTUFileHandler.VTUKeyWords

Base.write

VTUFileHandler.add_interpolation_keywords

VTUFileHandler.add_uncompress_keywords

VTUFileHandler.set_interpolation_keywords

VTUFileHandler.set_uncompress_keywords

General functions

VTUFileHandler.VTUHeader — Type

VTUHeader(::Type{T},input::Vector{UInt8}) where {T<:Union{UInt32,UInt64}}

Computes the VTUHeader based on the headertype and a Base64 decoded input data array.

Constructor

::Type{T}: headertype, either UInt32 or UInt64

input::Vector{UInt8}: input data

Fields

num_blocks::T : number of blocks

blocksize::T : size of blocks

last_blocksize::T : size of last block (can be different)

compressed_blocksizes::T : size of compressed blocks

VTUFileHandler.VTUDataField — Type

VTUDataField(x::Vector{T}) where {T} = new{T}(x)

Container for VTU field data.

Constructor

x::Vector{T}: Vector with VTU field data

Fields

dat::Vector{T} : VTU field data

VTUFileHandler.VTUData — Type

VTUData(dataarrays::Vector{XMLElement},appendeddata::Vector{XMLElement},headertype::Union{Type{UInt32},Type{UInt64}},offsets::Vector{Int},compressed_dat::Bool)

Container for VTU data.

Constructor

dataarrays::Vector{XMLElement}: Vector with all XML elements with tag DataArray of VTU file

appendeddata::Vector{XMLElement}: Vector with all XML elements with tag AppendedData

headertype::Union{Type{UInt32},Type{UInt64}} : Type of VTU header

offsets::Vector{Int} : Offset of each field data in the compressed appended data

compressed_dat::Bool : True if data is compressed

Fields

names::Vector{String} : Name of each data field

header::Vector{VTUHeader} : Vector with VTUHeaders.

data::Vector{VTUDataField} : Vector with VTUDataFields

interp_data::Vector{VTUDataField{Float64}} : Vector with VTUDataFields

idat::Vector{Int} : Vector indexing the data-fields onto which the math operators should be applied

For a VTUDataField to appear in data, the appropriate keyword has to be added via add_uncompress_keywords or add_interpolation_keywords. For a VTUDataField to appear in interp_data the appropriate keyword has to be added via add_interpolation_keywords. For more information, see VTUKeyWords.

VTUFileHandler.VTUFile — Type

VTUFile(name::String)

Loads a VTU file. Don't forget to set the proper fieldnames via set_uncompress_keywords and set_interpolation_keywords.

Constructor

name::String: path to vtu file

Fields

name::String: path to vtu file; destination for file writing

xmlroot::XMLElement: VTU file in XML represantation

dataarrays::Vector{XMLElement}: Vector with all XML elements with tag DataArray of VTU file

headertype::Union{Type{UInt32},Type{UInt64}}: type of header

offsets::Vector{Int}: Offset of each field data in the compressed appended data

data::VTUData: Conatainer with VTUData

compressed_dat::Bool : True if data is compressed

Example

set_uncompress_keywords("temperature","points")
set_interpolation_keywords("temperature")
vtufile = VTUFile("./path-to-vtu/example.vtu");

VTU Keywords

VTUFileHandler.VTUKeyWords — Type

VTUKeyWords

Container for interpolation and uncompress keywords. Will be created as a module-global constant value vtukeywords.

Fields

interpolation_keywords::Vector{String} : Vector with names of VTU fields onto which the math operators should apply

uncompress_keywords : Vector with names of VTU field that should be uncompressed and stored in memory

Standard values

By default, keywords were set that match results of the simulation software OpenGeoSys.

interpolation_keywords = ["displacement","epsilon","pressure_interpolated","sigma","temperature_interpolated"]
uncompress_keywords = ["connectivity","offsets","bulk_node_ids","bulk_element_ids","Points","MaterialIDs"]

Example

If a VTU result file e.g. owns a field named field1 to which the math operators should be applied, it can be added via:

add_interpolation_keywords(["field1"]])

VTUFileHandler.set_uncompress_keywords — Function

set_uncompress_keywords(uk::Vector{String})

Sets all fields that should be accessible and therefore needs to be uncompressed

Arguments

uk::Vector{String}: fieldnames

VTUFileHandler.add_uncompress_keywords — Function

add_uncompress_keywords(uk::Vector{String})

Adds fields that should be accessible and therefore needs to be uncompressed

Arguments

uk::Vector{String}: fieldnames

VTUFileHandler.set_interpolation_keywords — Function

set_interpolation_keywords(ik::Vector{String})

Sets all fields onto which the math operators should be applied

Arguments

ik::Vector{String}: fieldnames

VTUFileHandler.add_interpolation_keywords — Function

add_interpolation_keywords(ik::Vector{String})

Adds fields onto which the math operators should be applied

Arguments

ik::Vector{String}: fieldnames

IO-Functions

Base.write — Method

write(vtufile::VTUFile,add_timestamp::Bool=true)

Writes a VTUFile to destination vtufile.name

Arguments

vtufile::VTUFile: VTU file

add_timestamp::Bool: adds a timestamp to vtufile.name if add_timestamp==true